feat: implementação de ordenação

+++ src/modules/categories/categories.service.ts
@@ -5,7 +5,9 @@ export class CategoriesService {
   constructor(private prisma: PrismaService) {}
 
   findAll() {
-    return this.prisma.category.findMany();
+    return this.prisma.category.findMany({
+      orderBy: [{ order: 'asc' }],
+    });
   }
 
   findById(id: string) {
@@ -14,12 +16,24 @@ export class CategoriesService {
     });
   }
 
-  create(data: { title: string; image?: string | null; isVisible?: boolean }) {
+  async create(data: {
+    title: string;
+    image?: string | null;
+    isVisible?: boolean;
+  }) {
+    const lastCategory = (await this.prisma.category.findFirst({
+      orderBy: { order: 'desc' },
+      select: { order: true },
+    })) as { order: number } | null;
+
+    const nextOrder = (lastCategory?.order ?? 0) + 1;
+
     return this.prisma.category.create({
       data: {
         title: data.title,
-        image: data.image,
-        isVisible: data.isVisible,
+        image: data.image ?? null,
+        isVisible: data.isVisible ?? true,
+        order: nextOrder,
       },
     });
   }
@@ -37,10 +51,46 @@ export class CategoriesService {
       },
     });
   }
+  async updateBatchOrder(items: { id: string; order: number }[]) {
+    return this.prisma.$transaction(
+      items.map((item) =>
+        this.prisma.category.update({
+          where: { id: item.id },
+          data: { order: item.order },
+        }),
+      ),
+    );
+  }
 
-  delete(id: string) {
-    return this.prisma.category.delete({
+  async delete(id: string) {
+    const category: { id: string; order: number } | null =
+      await this.prisma.category.findUnique({
+        where: { id },
+        select: {
+          id: true,
+          order: true,
+        },
+      });
+
+    if (!category) return null;
+
+    await this.prisma.category.delete({
       where: { id },
     });
+
+    await this.prisma.category.updateMany({
+      where: {
+        order: {
+          gt: category.order,
+        },
+      },
+      data: {
+        order: {
+          decrement: 1,
+        },
+      },
+    });
+
+    return { id: category.id };
   }
 }
